clean comments, move utils constants
